move drawing fxn to xlib/drawing.c and tidy up main.h a bit
